+++ ui/src/main/java/com/theoplayer/android/ui/SubtitleTrackList.kt
@@ -2,13 +2,11 @@ package com.theoplayer.android.ui
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.ListItem
 import androidx.compose.material3.RadioButton
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import java.util.*
 
 /**
  * A list of subtitle tracks, from which the user can choose an active subtitle track.
@@ -16,7 +14,6 @@ import java.util.*
  * @param modifier the [Modifier] to be applied to this menu
  * @param onClick called when a subtitle track in the list is clicked
  */
-@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun SubtitleTrackList(
     modifier: Modifier = Modifier,
@@ -28,7 +25,7 @@ fun SubtitleTrackList(
     LazyColumn(modifier = modifier) {
         item(key = null) {
             ListItem(
-                headlineText = { Text(text = "Off") },
+                headlineContent = { Text(text = "Off") },
                 leadingContent = {
                     RadioButton(
                         selected = (activeSubtitleTrack == null),
@@ -47,7 +44,7 @@ fun SubtitleTrackList(
         ) {
             val audioTrack = subtitleTracks[it]
             ListItem(
-                headlineText = { Text(text = formatTrackLabel(audioTrack)) },
+                headlineContent = { Text(text = formatTrackLabel(audioTrack)) },
                 leadingContent = {
                     RadioButton(
                         selected = (activeSubtitleTrack == audioTrack),